Decouple navbar into component
